requirements.txt
# Core ML / deep learning
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.49.0       # Qwen2.5-VL and general HuggingFace utilities
accelerate>=0.29.0         # multi-GPU / mixed-precision helpers
qwen_vl_utils

# Vision utilities
Pillow>=10.0.0
opencv-python-headless>=4.9.0

# Dataset / data processing
datasets>=2.18.0           # HuggingFace datasets
huggingface_hub>=0.22.0    # For downloading datasets and models directly from HF
numpy>=1.26.0
pandas>=2.2.0
nuscenes-devkit
dotenv

# Evaluation
pycocotools>=2.0.7         # for IOU calculations
scipy>=1.13.0              # ADE / trajectory metrics

# nuScenes dependencies (explicit to avoid devkit downgrade conflicts)
pyquaternion>=0.9.9
Shapely>=2.0.0
cachetools
fire
